% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adaptive.mcmc.cpp}
\alias{adaptive.mcmc.cpp}
\title{Adaptive MCMC algorithm implemented in C++}
\usage{
adaptive.mcmc.cpp(lprior, llikelihood, outfun, nburn, initial, nbatch,
  blen = 1L)
}
\arguments{
\item{lprior}{A function returning the log prior probability of the parameters}

\item{llikelihood}{A function returning the log likelihood of the parameters given the data}

\item{outfun}{A function that is called with the current parameters. It's output is added to the returned list}

\item{nburn}{Number of iterations of burn in}

\item{initial}{Vector with starting parameter values}

\item{nbatch}{Number of batches to run (number of samples to return)}

\item{blen}{Length of each batch}
}
\value{
Returns a list with the accepted samples and the corresponding llikelihood values
}
\description{
MCMC which adapts its proposal distribution for faster convergence following:
Sherlock, C., Fearnhead, P. and Roberts, G.O. The Random Walk Metrolopois: Linking Theory and Practice Through a Case Study. Statistical Science 25, no.2 (2010): 172-190.
}
\seealso{
\code{\link{adaptive.mcmc}} For a more flexible R frontend to this function.
}

