% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{age_group_limits}
\alias{age_group_limits}
\title{Extract upper age group limits from age group level description}
\usage{
age_group_limits(levels)
}
\arguments{
\item{levels}{The levels representing each age groups.}
}
\value{
Vector representing the age group(s) limits
}
\description{
Returns a vector of age group limits given the age group level descriptions. This is a helper function, which is essentially the reverse of \code{\link{age_group_levels}}.
}
\seealso{
\code{\link{age_group_levels}} For the reverse of this function.
}
