% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{age_sizes}
\alias{age_sizes}
\title{Number of people of each age for the UK in 1999}
\format{\preformatted{'data.frame':	86 obs. of  1 variable:
 $ V1: int  625100 632100 648500 639800 645800 663500 668700 690300 700100 688200 ...
}}
\usage{
age_sizes
}
\description{
Number of people of each age for the UK in 1999
}
\keyword{datasets}

