% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{aggregateModel}
\alias{aggregateModel}
\title{Aggregate model results at different time points}
\usage{
aggregateModel(func, batch, aggregate, ...)
}
\arguments{
\item{func}{The function that gets called for each set of parameters (e.g. infectionODEs)}

\item{batch}{Posterior parameters samples resulting from mcmc. Each row is a set of parameters}

\item{aggregate}{The aggragation function (i.e. mean, var etc). If the function returns a list, then the names are used in the return value}

\item{...}{Extra parameters passed to func.}
}
\value{
Returns a data.frame in long format, with for each time point, interval, column.ID the value
}
\description{
This function is useful to convert the mcmc results into aggregated results, such as mean, variance etc.
}

