% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\name{as.vaccination.calendar}
\alias{as.vaccination.calendar}
\title{Create a vaccination calendar object}
\usage{
as.vaccination.calendar(legacy = NULL, no.risk.groups = NULL,
  no.age.groups = NULL)
}
\arguments{
\item{legacy}{An optional legacy object to convert. Will sanity check the given object}

\item{no.risk.groups}{The total number of risk groups (optional)}

\item{no.age.groups}{The total number of age groups (optional)}
}
\value{
A list that contains the \code{calendar} and \code{efficacy} of the vaccine for that year
}
\description{
Helper function to create (valid) vaccination calendars. It will take the given data, do some
sanity checking and convert it into a list object that can be passed to other functions.
}

