% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_R0}
\alias{as_R0}
\title{Calculate R0 from transmission rate}
\usage{
as_R0(transmission_rate, contact_matrix, age_groups, duration = 1.8)
}
\arguments{
\item{transmission_rate}{The transmission rate of the disease}

\item{age_groups}{The population size of the different age groups}

\item{duration}{Duration of the infectious period. Default value is 1.8 days}

\item{contaxt_matrix}{The contact matrix between age groups}
}
\value{
Returns the R0
}
\description{
Uses the transmission rate (\eqn{\lambda}), contact matrix (\eqn{c}), population (\eqn{N}), and infectious period (\eqn{\gamma}) 
to calculate the R0 using the following equation.
\deqn{\lambda max(EV(C)) \gamma}
where \eqn{EV(C)} denotes the eigenvalues of the matrix \eqn{C} which is calculated from the contact matrix and the population 
(\eqn{C[i,j] = c[i,j] N[j]}).
}
