% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_transmission_rate}
\alias{as_transmission_rate}
\title{Calculate transmission rate from R0}
\usage{
as_transmission_rate(R0, contact_matrix, age_groups, duration = 1.8)
}
\arguments{
\item{R0}{The R0 of the disease}

\item{contact_matrix}{The contact matrix between age groups}

\item{age_groups}{The population size of the different age groups}

\item{duration}{Duration of the infectious period. Default value is 1.8 days}
}
\value{
Returns the transmission rate
}
\description{
Uses the R0 (\eqn{R0}), contact matrix (\eqn{c}), population (\eqn{N}), and infectious period (\eqn{\gamma}) 
to calculate the transmission rate using the following equation.
\deqn{R0/(max(EV(C)) \gamma)}
where \eqn{EV(C)} denotes the eigenvalues of the matrix \eqn{C} which is calculated from the contact matrix and the population 
(\eqn{C[i,j] = c[i,j] N[j]}).
}
