% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binomial.inference}
\alias{binomial.inference}
\title{MCMC based inference of the parameter values given the different data sets}
\usage{
binomial.inference(age_sizes, ili, mon_pop, n_pos, n_samples, vaccine_calendar,
  polymod_data, initial, nburn = 10000L, nbatch = 1000L, blen = 1L)
}
\arguments{
\item{age_sizes}{A vector with the population size by each age {1,2,..}}

\item{ili}{The number of Influenza-like illness cases per week}

\item{mon_pop}{The number of people monitored for ili}

\item{n_pos}{The number of positive samples for the given strain (per week)}

\item{n_samples}{The total number of samples tested}

\item{vaccine_calendar}{A vaccine calendar valid for that year}

\item{polymod_data}{Contact data for different age groups}

\item{initial}{Vector with starting parameter values}

\item{nburn}{Number of iterations of burn in}

\item{nbatch}{Number of batches to run (number of samples to return)}

\item{blen}{Length of each batch}
}
\value{
Returns a list with the accepted samples and the corresponding llikelihood values
}
\description{
Based on a simplified (binomial) log likelihood function, that does not
assume the confirmation samples to be a strict subsample of the ILI cases
}

