% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{confirmed.samples}
\alias{confirmed.samples}
\title{Number of confirmed positive samples per week}
\format{A list with both the number of positive samples and the total number samples taken
\describe{
  \item{positive}{Matrix with total number of positive samples per week (row) and age group (column)}
  \item{total.samples}{The total number of samples tested over the same time/age group}
}}
\usage{
confirmed.samples
}
\description{
Generally only for a fraction of the Influenza-like illness cases an actual (blood) sample is taken to confirm whether/which strain the subject is infected with.
}
\keyword{datasets}

