% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{contact_matrix}
\alias{contact_matrix}
\title{Create a contact matrix based on polymod data.}
\usage{
contact_matrix(polymod_data, demography, age_group_limits = as.numeric(c(1, 5,
  15, 25, 45, 65)))
}
\arguments{
\item{polymod_data}{Contact data for different age groups}

\item{demography}{A vector with the population size by each age {0,1,2,..}}

\item{age_group_limits}{The upper limits of the different age groups (by default: c(1,5,15,25,45,65), which corresponds to age groups: <1, 1-14, 15-24, 25-44, 45-64, >=65.}
}
\value{
Returns a symmetric matrix with the frequency of contact between each age group
}
\description{
Create a contact matrix based on polymod data.
}
