% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{coverage}
\alias{coverage}
\title{Data on vaccine coverage during the 2007-08 season}
\format{An object of class \code{data.frame} with 19 rows and 6 columns.}
\usage{
coverage
}
\description{
Data on vaccine coverage during the 2007-08 season
}
\keyword{datasets}

