% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{credible.interval.model}
\alias{credible.interval.model}
\title{Calculate the credible interval at different time points}
\usage{
credible.interval.model(func, batch, intervals = c(0.01, 0.5, 0.99), ...)
}
\arguments{
\item{func}{The function that gets called for each set of parameters (e.g. infectionODEs)}

\item{batch}{Posterior parameters samples resulting from mcmc. Each row is a set of parameters}

\item{intervals}{The cut offs at which to calculate the intervals}

\item{...}{Extra parameters passed to func.}
}
\value{
Returns a data.table in long format, with for each time point, interval, column.ID the value
}
\description{
This function is useful to convert the mcmc results into credible intervals, which is needed for plotting
your results.
}

