% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demography}
\alias{demography}
\title{Demography of the UK in 1999}
\format{An object of class \code{integer} of length 86.}
\usage{
demography
}
\description{
A vector containing the number of people of a certain age in the UK. The first element
is people of age 0, second is age 1 etc.
}
\keyword{datasets}
