% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmultinom.cpp}
\alias{dmultinom.cpp}
\title{Probability density function for multinomial distribution}
\usage{
dmultinom.cpp(x, size, prob, use_log = FALSE)
}
\arguments{
\item{x}{The counts}

\item{size}{The total size from which is being samples}

\item{prob}{Probabilities of each different outcome}

\item{use_log}{Whether to return logarithm probability}
}
\value{
The probability of getting the counts, given the total size and probability of drawing each.
}
\description{
Probability density function for multinomial distribution
}

