% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTimeFromWeekYear}
\alias{getTimeFromWeekYear}
\title{Convert given week in given year into an exact date corresponding to the Monday of that week}
\usage{
getTimeFromWeekYear(week, year)
}
\arguments{
\item{week}{The number of the week we need the date of}

\item{year}{The year}
}
\value{
The date of the Monday in that week
}
\description{
Convert given week in given year into an exact date corresponding to the Monday of that week
}

