% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_effectiveness.R
\name{hospitalisations}
\alias{hospitalisations}
\title{Calculate number of hospitalisations from incidence data}
\usage{
hospitalisations(proportion, incidence, no_risk_groups = NULL,
  no_age_groups = NULL)
}
\arguments{
\item{proportion}{The expected proportion of cases resulting in hospitalisations. 
This can be a constant value for each age/risk groups, or a vector with the proportion for each age/risk group. Finally if
either no_risk_groups or no_age_groups is specified it can also be a vector with the proportion by risk groups or by age group.}

\item{incidence}{Vector containing the expected incidence for each age/risk group as returned by for example \code{\link{vaccinationScenario}}.}

\item{no_risk_groups}{The total number of risk groups (optional)}

\item{no_age_groups}{The total number of age groups (optional)}
}
\value{
A vector with the number of hospitalisations for each age/risk group
}
\description{
Uses the provided proportion to calculate the number of hosptitalisations from the incidence data. The expected
proportion of hospitalisations needs to be provided by the user and is likely to be country/disease specific.
}

