% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{ili}
\alias{ili}
\title{Number of Influenza-like illness cases per week}
\format{A list with both the number of cases and the total number of monitored people
\describe{
  \item{ili}{Matrix with total number of cases per week (row) and age group (column)}
  \item{total.monitored}{The number of people monitored over the same time/age group}
}}
\usage{
ili
}
\description{
Number of Influenza-like illness cases per week
}
\keyword{datasets}
