% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infection.model}
\alias{infection.model}
\title{Run the SEIR model for the given parameters}
\usage{
infection.model(age_sizes, vaccine_calendar, polymod_data, susceptibility,
  transmissibility, init_pop, infection_delays, interval = 1L)
}
\arguments{
\item{age_sizes}{A vector with the population size by each age {1,2,..}}

\item{vaccine_calendar}{A vaccine calendar valid for that year}

\item{polymod_data}{Contact data for different age groups}

\item{susceptibility}{Vector with susceptibilities of each age group}

\item{transmissibility}{The transmissibility of the strain}

\item{init_pop}{The (log of) initial infected population}

\item{infection_delays}{Vector with the time of latent infection and time infectious}

\item{interval}{Interval (in days) between data points}
}
\value{
A data frame with number of new cases after each interval during the year
}
\description{
Run the SEIR model for the given parameters
}

