% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infectionODEs.cpp}
\alias{infectionODEs.cpp}
\title{Run the SEIR model for the given parameters}
\usage{
infectionODEs.cpp(population, initial_infected, vaccine_calendar,
  contact_matrix, susceptibility, transmissibility, infection_delays, dates)
}
\arguments{
\item{population}{The population size of the different age groups, subdivided into risk groups}

\item{initial_infected}{The corresponding number of initially infected}

\item{vaccine_calendar}{A vaccine calendar valid for that year}

\item{contact_matrix}{Contact rates between different age groups}

\item{susceptibility}{Vector with susceptibilities of each age group}

\item{transmissibility}{The transmissibility of the strain}

\item{infection_delays}{Vector with the time of latent infection and time infectious}

\item{interval}{Interval (in days) between data points}
}
\value{
A data frame with number of new cases after each interval during the year
}
\description{
Run the SEIR model for the given parameters
}

