% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{inference}
\alias{inference}
\title{MCMC based inference of the parameter values given the different data sets}
\usage{
inference(demography, ili, mon_pop, n_pos, n_samples, vaccine_calendar,
  polymod_data, initial, parameter_map, age_group_map, risk_group_map,
  risk_ratios, nburn = 0, nbatch = 1000, blen = 1)
}
\arguments{
\item{demography}{A vector with the population size by each age {0,1,..}}

\item{ili}{The number of Influenza-like illness cases per week}

\item{mon_pop}{The number of people monitored for ili}

\item{n_pos}{The number of positive samples for the given strain (per week)}

\item{n_samples}{The total number of samples tested}

\item{vaccine_calendar}{A vaccine calendar valid for that year}

\item{polymod_data}{Contact data for different age groups}

\item{initial}{Vector with starting parameter values}

\item{parameter_map}{Optional mapping parameter (by description and age group) to the relevant index
in the initial vector. Needed parameters are: epsilon (ascertainmen) with a separater value per data
age group, transmissibility, psi (infection from outside sources), susceptibility (with a value per age group)
and log of initial_infected population.}

\item{age_group_map}{Optional age group mapping from model age groups to data age groups (\code{\link{age_group_mapping}})}

\item{risk_group_map}{Optional risk group mapping from model risk groups to data risk groups (\code{\link{risk_group_mapping}})}

\item{nburn}{Number of iterations of burn in}

\item{nbatch}{Number of batches to run (number of samples to return)}

\item{blen}{Length of each batch}
}
\value{
Returns a list with the accepted samples and the corresponding llikelihood values and a matrix (contact.ids) containing the ids (row number) of the contacts data used to build the contact matrix.
}
\description{
MCMC based inference of the parameter values given the different data sets
}
