% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inference.results}
\alias{inference.results}
\title{MCMC Result produced by the inference function}
\format{An object of class \code{list} of length 3.}
\usage{
inference.results
}
\description{
MCMC Result produced by the inference function
}
\keyword{datasets}
