% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inference_multistrains}
\alias{inference_multistrains}
\title{MCMC based inference of the parameter values given the different data sets based on multiple strains}
\usage{
inference_multistrains(age_sizes, ili, mon_pop, n_pos, n_samples,
  vaccine_calendar, polymod_data, initial, nburn = 0L, nbatch = 1000L,
  blen = 1L)
}
\arguments{
\item{age_sizes}{A vector with the population size by each age {1,2,..}}

\item{ili}{The number of Influenza-like illness cases per week}

\item{mon_pop}{The number of people monitored for ili}

\item{n_pos}{The number of positive samples per strain (per week)}

\item{n_samples}{The total number of samples tested}

\item{vaccine_calendar}{Vaccine calendars per strain valid for that year}

\item{polymod_data}{Contact data for different age groups}

\item{initial}{Vector with starting parameter values}

\item{nburn}{Number of iterations of burn in}

\item{nbatch}{Number of batches to run (number of samples to return)}

\item{blen}{Length of each batch}
}
\value{
Returns a list with the accepted samples and the corresponding llikelihood values and a matrix (contact.ids) containing the ids (row number) of the contacts data used to build the contact matrix.
}
\description{
MCMC based inference of the parameter values given the different data sets based on multiple strains
}

