% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llikelihood.cases}
\alias{llikelihood.cases}
\title{Returns log likelihood of the predicted number of cases given the data for that week}
\usage{
llikelihood.cases(epsilon, psi, predicted, population_size, ili_cases,
  ili_monitored, confirmed_positive, confirmed_samples)
}
\arguments{
\item{epsilon}{Parameter for the probability distribution}

\item{psi}{Parameter for the probability distribution}

\item{predicted}{Number of cases predicted by your model}

\item{population_size}{The total population size in the relevant age group}

\item{ili_cases}{The number of Influenza Like Illness cases}

\item{ili_monitored}{The size of the population monitored for ILI}

\item{confirmed_positive}{The number of samples positive for the Influenza strain}

\item{confirmed_samples}{Number of samples tested for the Influenza strain}
}
\description{
The model results in a prediction for the number of new cases in a certain age group and for a certain week. This function calculates the likelihood of that given the data on reported Influenza Like Illnesses and confirmed samples.
}

