% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_likelihood_cases}
\alias{log_likelihood_cases}
\title{Returns log likelihood of the predicted number of cases given the data}
\usage{
log_likelihood_cases(epsilon, psi, predicted, population_size, ili_cases,
  ili_monitored, confirmed_positive, confirmed_samples, depth = 2L)
}
\arguments{
\item{epsilon}{Parameter for the probability distribution by age group}

\item{psi}{Parameter for the probability distribution}

\item{predicted}{Number of cases predicted by your model for each week and age group}

\item{population_size}{The total population size in the age groups}

\item{ili_cases}{The number of Influenza Like Illness cases by week and age group}

\item{ili_monitored}{The size of the population monitored for ILI  by week and age group}

\item{confirmed_positive}{The number of samples positive for the Influenza strain  by week and age group}

\item{confirmed_samples}{Number of samples tested for the Influenza strain  by week and age group}

\item{depth}{Depth/precision of the approximation. In general the a value of 2 is used. Higher is more precise.}
}
\description{
The model results in a prediction for the number of new cases in a certain age group and for a certain week. This function sum the log likelihood for the predicted cases for each week and age group given the data on reported Influenza Like Illnesses and confirmed samples.
}
