% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{mcmcsample}
\alias{mcmcsample}
\title{MCMC Result produced by the inference function}
\format{\preformatted{List of 4
 $ parameters     :List of 5
  ..$ epsilon         : num [1:5] 0.0119 0.0119 0.0183 0.0183 0.0543
  ..$ psi             : num 1.05e-05
  ..$ transmissibility: num 0.166
  ..$ susceptibility  : num [1:7] 0.386 0.386 0.386 0.927 0.927 ...
  ..$ init_pop        : num -0.154
 $ time_infectious: num 1.8
 $ time_latent    : num 0.8
 $ contact_ids    : num [1:597] 595 480 301 424 10 208 333 486 333 465 ...
}}
\usage{
mcmcsample
}
\description{
MCMC Result produced by the inference function
}
\keyword{datasets}

