% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_effectiveness.R
\name{mortality}
\alias{mortality}
\title{Calculate number of deaths from incidence data}
\usage{
mortality(proportion, incidence, no_risk_groups = NULL,
  no_age_groups = NULL)
}
\arguments{
\item{proportion}{The expected proportion of cases resulting in death. 
This can be a constant value for each age/risk groups, or a vector with the proportion for each age/risk group. Finally if
either no_risk_groups or no_age_groups is specified it can also be a vector with the proportion by risk groups or by age group.}

\item{incidence}{Vector containing the expected incidence for each age/risk group as returned by for example \code{\link{vaccinationScenario}}.}

\item{no_risk_groups}{The total number of risk groups (optional)}

\item{no_age_groups}{The total number of age groups (optional)}
}
\value{
A vector with the number of deaths for each age/risk group
}

