% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodel.R
\name{parameter_mapping}
\alias{parameter_mapping}
\title{Mapping parameters to the model}
\usage{
parameter_mapping(epsilon, psi, transmissibility, susceptibility,
  initial_infected, parameters)
}
\arguments{
\item{epsilon}{A vector holding the indeces of the ascertainment parameters for each age group in the parameters vector}

\item{psi}{Index of the psi parameter in the parameters vector}

\item{transmissibility}{Index of the transmissibility parameter in the parameters vector}

\item{susceptibility}{A vector holding the indeces of the susceptibility parameter for each age group in the parameters vector}

\item{initial_infected}{Index of the initial_infected parameter in the parameters vector}

\item{parameters}{Optional parameters vector. For simple cases it is possible to infer the parameter map just from the (named) parameters vector. In that case
no other variables need to be passed to this function.}
}
\value{
A list specifying the parameter mapping as used in \code{\link{inference}} and 
\code{\link{vaccination_scenario}}
}
\description{
To reduce complexity it is common to map certain parameters to multiple age groups. For example
in the UK model we use the same susceptibility for age groups 1,2 and 3. Similarly for fitting purposes we
need a vector of parameters. That vector will hold different parameters: ascertainment (epsilon; by age group), 
psi, transmissibility, susceptibility (by age group) and initial infected. This method makes this easier.

In the UK model we use 9 parameters, the first three are the epsilon for age groups 1,2 and 3,4 and 5. This is then followed by the 
psi parameter (4th parameter in the parameter vector). Next is the transmissibility parameter. Then three susceptibility parameters, used in
age groups: 1,2,3 and 4,5,6 and 7. Finally followed by initial_infected as the 9th parameter in the parameters vector. 
\code{parameter_map(epsilon = c(1,1,2,2,3), psi = 4, transmissibility = 5, susceptibility = c(6,6,6,7,7,7,8), initial_infected = 9}
}
