% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polymod_uk}
\alias{polymod_uk}
\title{Polymod contact data for the UK}
\format{Data frame containing all polymod data for the UK
\describe{
   \item{Age}{Age of the subject}
   \item{Weekend}{Data taken in the weekend (1) or during the week (0)}
   \item{...}{Number of contacts with individuals from the age groups}
}}
\usage{
polymod_uk
}
\description{
Polymod contact data for the UK
}
\keyword{datasets}
