% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\name{read.legacy.vaccine.file}
\alias{read.legacy.vaccine.file}
\title{Function to read legacy file format for vaccine calendar}
\usage{
read.legacy.vaccine.file(file)
}
\arguments{
\item{file}{The file to load}
}
\value{
A list that contains the \code{calendar} and \code{efficacy} of the vaccine for that year
}
\description{
Function to read legacy file format for vaccine calendar
}
