% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_ext.R
\name{stat_ci}
\alias{stat_ci}
\title{Draw credibility intervals}
\usage{
stat_ci(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Will draw credibitlity intervals with different colours for each interval
}

