% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_ext.R
\docType{data}
\name{StatCI}
\alias{StatCI}
\title{Stat object for drawing credible intervals}
\format{An object of class \code{StatCI} (inherits from \code{Stat}, \code{ggproto}) of length 5.}
\usage{
StatCI
}
\description{
Converts the data to show a polygon
}
\keyword{datasets}

