% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stratify_by_age}
\alias{stratify_by_age}
\title{Stratify the population by age}
\usage{
stratify_by_age(age_sizes, limits = as.numeric(c(1, 5, 15, 25, 45, 65)))
}
\arguments{
\item{age_sizes}{A vector containing the population size by age (first element is number of people of age 1 and below)}

\item{limits}{The upper limit to each age groups (not included) (1,5,15,25,45,65) corresponds to the following age groups: <1, 1-4, 5-14, 15-24, 25-44, 45-64 and >=65.}
}
\value{
A vector with the population in each age group.
}
\description{
Stratifies the population and returns the population size of each age group.
}

