% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{uptake}
\alias{uptake}
\title{Data on the uptake rates of vaccination in the UK}
\format{\preformatted{'data.frame':	14 obs. of  8 variables:
 $ Year              : int  199596 199697 199798 199899 199900 200001 200102 200203 200304 200405 ...
 $ under.65.1st.month: num  61.7 68.7 72.1 68.1 66.3 ...
 $ over.65.1st.month : num  70.3 76.6 79.4 75.7 74.2 ...
 $ under.65.2nd.month: num  88.9 92 93.8 92.2 91.7 ...
 $ over.65.2nd.month : num  93.3 95.8 96.7 95.7 94.9 ...
 $ under.65.3rd.month: num  98.3 98.1 98.7 96.9 97.2 ...
 $ over.65.3rd.month : num  99.2 99.3 99.8 99.2 98.9 ...
 $ NA                : chr  "Joseph et al." "Joseph et al." "Joseph et al." "Joseph et al." ...
}}
\usage{
uptake
}
\description{
Data on the uptake rates of vaccination in the UK
}
\keyword{datasets}

