% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\name{vaccination_scenario}
\alias{vaccination_scenario}
\title{Calculate number of influenza cases given a vaccination strategy}
\usage{
vaccination_scenario(vaccine_calendar, parameters, contact_ids,
  incidence_function, time_column, parameter_map, ..., verbose = T)
}
\arguments{
\item{vaccine_calendar}{A vaccine calendar valid for that year}

\item{parameters}{The parameters to use. Both a vector or a data frame with each row a set of parameters 
(e.g. a batch of inferred parameters by adaptive_mcmc$batch) are accepted.}

\item{contact_ids}{Optional: The contact_ids used to infer the contact matrix. Similar to the \code{parameters} this can be a
vector or a data frame.}

\item{incidence_function}{An optional function that takes a \code{vaccine_calendar}, \code{parameters} and optionally
\code{contact_ids} and returns the incidence over time. If none is provided then the default
\code{infectionODEs} is used. In that case both \code{polymod_data} and \code{demography} data need to be
provided as well.}

\item{time_column}{An optional time column that will be removed from the data returned by the incidence_function}

\item{parameter_map}{Optional: mapping for the parameters (by description and age group) to the relevant index
in the parameter vector. Needed parameters are: transmissibility, psi (infection from outside sources), susceptibility (with a value per age group)
and log of initial_infected population \code{\link{parameter_mapping}}.}

\item{...}{Further parameters that will be passed to the \code{incidence_function}}

\item{verbose}{Whether to display warnings. Default is TRUE.}
}
\value{
The total incidence in a year per age and risk group
}
\description{
Calculate number of influenza cases given a vaccination strategy
}
\seealso{
\code{\link{infectionODEs}}
}
