% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vaccinationScenario}
\alias{vaccinationScenario}
\title{Calculate number of influenza cases given a vaccination strategy}
\usage{
vaccinationScenario(age_sizes, vaccine_calendar, polymod_data, contact_ids,
  parameters)
}
\arguments{
\item{age_sizes}{A vector with the population size by each age {1,2,..}}

\item{vaccine_calendar}{A vaccine calendar valid for that year}

\item{polymod_data}{Contact data for different age groups}

\item{contact_ids}{IDs (row numbers) of the contact data used when modelling this scenario}

\item{parameters}{The parameters to use}
}
\value{
A data frame with the total number of influenza cases in that year
}
\description{
Superseded by \code{vaccination_scenario}
}
\seealso{
\code{\link{vaccination_scenario}}
}
\keyword{internal}
