% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{vaccine.calendar}
\alias{vaccine.calendar}
\title{Function to create a vaccine calendar object}
\usage{
vaccine.calendar(coverage, efficacy, uptake)
}
\arguments{
\item{coverage}{A vector containing the total coverage for the vaccine for each age/risk group}

\item{efficacy}{The efficacy of the vaccine for that year (subdivided by each age group)}

\item{uptake}{The uptake of the vaccine over different time periods}
}
\value{
A list that contains the \code{calendar} and \code{efficacy} of the vaccine for that year
}
\description{
Deprecated
}
\keyword{internal}
