% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{vaccine_calendar}
\alias{vaccine_calendar}
\title{Vaccination uptake rate for the UK in 1999}
\format{\preformatted{List of 2
 $ efficacy: num [1:7] 0.7 0.7 0.7 0.7 0.7 0.7 0.46
 $ calendar: num [1:123, 1:21] 1.19e-05 1.19e-05 1.19e-05 1.19e-05 1.19e-05 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:21] "V1" "V2" "V3" "V4" ...
}}
\usage{
vaccine_calendar
}
\description{
Vaccination uptake rate for the UK in 1999
}
\keyword{datasets}

