% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_effectiveness.R
\name{vaccine_doses}
\alias{vaccine_doses}
\title{Calculate number of vaccine doses needed based on vaccine calendar}
\usage{
vaccine_doses(vaccination_calendar, population)
}
\arguments{
\item{vaccination_calendar}{The vaccination calendar. This object should have the same layout as a 
vaccination_calendar returned by \code{\link{as.vaccination.calendar}}.}

\item{population}{A vector with the size of the population in each age and risk group}
}
\value{
The number of doses by age and risk group.
}
\description{
Calculates the number of doses required to vaccinate the population under the provided vaccination calendar.
}

