% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{age_sizes}
\alias{age_sizes}
\title{Number of people of each age for the UK in 1999}
\format{An object of class \code{data.frame} with 86 rows and 1 columns.}
\usage{
age_sizes
}
\description{
Number of people of each age for the UK in 1999
}
\keyword{internal}
