% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\name{as.vaccination.calendar}
\alias{as.vaccination.calendar}
\title{Create a vaccination calendar object}
\usage{
as.vaccination.calendar(efficacy = NULL, dates = NULL, coverage = NULL,
  legacy = NULL, no_risk_groups = NULL, no_age_groups = NULL,
  starting_year = NULL)
}
\arguments{
\item{efficacy}{Vector holding the vaccine efficacy for each age group and risk group}

\item{dates}{Vector containing the dates at which the coverage was measured.}

\item{coverage}{Data frame with the fraction of coverage for each age and risk group in each column. Different rows are the different dates of measurements}

\item{legacy}{An optional legacy object to convert. When passed a legacy object this function will sanity check the given object}

\item{no_risk_groups}{The total number of risk groups (optional)}

\item{no_age_groups}{The total number of age groups (optional)}

\item{starting_year}{The year of the start of the season. Only used when passed a legacy file}
}
\value{
A list that contains the \code{calendar} and \code{efficacy} of the vaccine for that year
}
\description{
Helper function to create (valid) vaccination calendars. It will take the given data, do some
sanity checking and convert it into a list object that can be passed to other functions.
}

