% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inference}
\alias{inference}
\title{MCMC based inference of the parameter values given the different data sets}
\usage{
inference(age_sizes, ili, mon_pop, n_pos, n_samples, vaccine_calendar,
  polymod_data, init_sample, mcmc_chain_length = 100000L, burn_in = 10000L,
  thinning = 100L)
}
\arguments{
\item{age_sizes}{A vector with the population size by each age {1,2,..}}

\item{ili}{The number of Influenza-like illness cases per week}

\item{mon_pop}{The number of people monitored for ili}

\item{n_pos}{The number of positive samples for the given strain (per week)}

\item{n_samples}{The total number of samples tested}

\item{vaccine_calendar}{A vaccine calendar valid for that year}

\item{polymod_data}{Contact data for different age groups}

\item{init_sample}{The initial parameters needed to run the ODE model (typically one of the posterior sample created when running the inference)}

\item{mcmc_chain_length}{The number of MCMC steps to sample from}

\item{burn_in}{The number of initial samples to skip}

\item{thinning}{Keep every so many samples}
}
\value{
A vector with posterior samples of the parameters (length of \code{mcmc_chain_length}/\code{thinning})
}
\description{
MCMC based inference of the parameter values given the different data sets
}

