% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{mcmcsample}
\alias{mcmcsample}
\title{DEPRECATED: MCMC Result produced by the inference function}
\format{An object of class \code{list} of length 4.}
\usage{
mcmcsample
}
\description{
DEPRECATED: MCMC Result produced by the inference function
}
\keyword{internal}
