% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{polymod_uk}
\alias{polymod_uk}
\title{Polymod contact data for the UK}
\format{Data frame containing all polymod data for the UK
\describe{
   \item{V1}{Age of the subject}
   \item{V2}{Data taken in the weekend (1) or during the week (0)}
   \item{V3-V9}{Number of contacts with persons from the different age groups}
}}
\usage{
polymod_uk
}
\description{
Polymod contact data for the UK
}
\keyword{datasets}

