% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_ext.R
\name{stat_ci}
\alias{stat_ci}
\title{Draw credibility intervals}
\usage{
stat_ci(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this layer}

\item{na.rm}{If FALSE (the default), removes missing values with a warning. If TRUE silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders.}

\item{...}{other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.}
}
\description{
Will draw credibitlity intervals with different colours for each interval
}
