% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{uptake}
\alias{uptake}
\title{Data on the uptake rates of vaccination in the UK}
\format{An object of class \code{data.frame} with 14 rows and 8 columns.}
\usage{
uptake
}
\description{
Data on the uptake rates of vaccination in the UK
}
\keyword{internal}
