% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vaccinationScenario}
\alias{vaccinationScenario}
\title{Calculate number of influenza cases given a vaccination strategy}
\usage{
vaccinationScenario(age_sizes, vaccine_calendar, polymod_data, sample)
}
\arguments{
\item{age_sizes}{A vector with the population size by each age {1,2,..}}

\item{vaccine_calendar}{A vaccine calendar valid for that year}

\item{polymod_data}{Contact data for different age groups}

\item{sample}{The parameters needed to run the ODE model (typically one of the posterior sample created when running the inference)}
}
\value{
A data frame with the total number of influenza cases in that year
}
\description{
Calculate number of influenza cases given a vaccination strategy
}

