% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{vaccine_calendar}
\alias{vaccine_calendar}
\title{Vaccination uptake rate for the UK in 1999}
\format{An object of class \code{list} of length 2.}
\usage{
vaccine_calendar
}
\description{
Vaccination uptake rate for the UK in 1999
}
\keyword{datasets}
